Refactor share group code
