+++ server/src/main/java/org/apache/kafka/server/share/ShareAcknowledgementBatch.java
@@ -1,0 +1,57 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.kafka.server.share;
+
+import java.util.List;
+
+/**
+ * The ShareAcknowledgementBatch containing the fields required to acknowledge the fetched records.
+ */
+public class ShareAcknowledgementBatch {
+
+  private final long firstOffset;
+  private final long lastOffset;
+  private final List<Byte> acknowledgeTypes;
+
+  public ShareAcknowledgementBatch(long firstOffset, long lastOffset, List<Byte> acknowledgeTypes) {
+    this.firstOffset = firstOffset;
+    this.lastOffset = lastOffset;
+    this.acknowledgeTypes = acknowledgeTypes;
+  }
+
+  public long firstOffset() {
+    return firstOffset;
+  }
+
+  public long lastOffset() {
+    return lastOffset;
+  }
+
+  public List<Byte> acknowledgeTypes() {
+    return acknowledgeTypes;
+  }
+
+  @Override
+  public String toString() {
+    return "ShareAcknowledgementBatch(" +
+        " firstOffset=" + firstOffset +
+        ", lastOffset=" + lastOffset +
+        ", acknowledgeTypes=" + ((acknowledgeTypes == null) ? "" : acknowledgeTypes) +
+        ")";
+  }
+}
